+++ tests/testthat/test-percentVJ.R
@@ -1,2 +1,0 @@
-# test script for percentVJ.R - testcases are NOT comprehensive!
-