+++ tests/testthat/test-vizGenes.R
@@ -1,1 +1,0 @@
-# test script for vizGenes.R - testcases are NOT comprehensive!